+++ run_tests.py
@@ -470,6 +470,8 @@ def main():
                            action="store_true", default=False)
     argparser.add_argument("--project", help="DNAnexus project ID",
                            default="project-F07pBj80ZvgfzQK28j35Gj54")
+    argparser.add_argument("--runtime-debug-level",
+                           help="printing verbosity of task/workflow runner, {0,1,2}")
     argparser.add_argument("--test", help="Run a test, or a subgroup of tests",
                            action="append", default=[])
     argparser.add_argument("--test-list", help="Print a list of available tests",
@@ -523,6 +525,9 @@ def main():
         compiler_flags.append("-force")
     if args.verbose:
         compiler_flags.append("-verbose")
+    if args.runtime_debug_level:
+        compiler_flags += ["-runtimeDebugLevel", args.runtime_debug_level]
+
     compiler_flags += ["--extras", os.path.join(top_dir, "test/extras.json")]
 
     if "call_native" in test_names: